+++ build.js
@@ -26,7 +26,10 @@ if (shell.exec(TSC_CMD).code !== 0) {
   shell.echo(chalk.red(`Error: Compiling ES6 modules failed`));
   shell.exit(1);
 }
+shell.mv(`${SRC_DIR}/indexForReactRouter.js`, `${SRC_DIR}/index.js`);
 shell.cp(`-Rf`, `${TYPES_DIR}/*`, SRC_DIR);
+shell.rm(`${TYPES_DIR}/indexForReactRouter.d.ts`);
+shell.mv(`${SRC_DIR}/indexForReactRouter.d.ts`, `${SRC_DIR}/index.d.ts`);
 
 // Maintain banners
 if (shell.exec(BANNER_CMD).code !== 0) {