+++ test/e2e/harness/src/App.tsx
@@ -13,8 +13,7 @@
 import * as React from 'react';
 import { Route, Switch, useHistory } from 'react-router-dom';
 import { OktaAuth } from '@okta/okta-auth-js';
-import { Security, LoginCallback, useOktaAuth } from '@okta/okta-react';
-import SecureRoute from '@okta/okta-react/src/SecureRoute';
+import { Security, LoginCallback, SecureRoute } from '@okta/okta-react/react-router';
 import Home from './Home';
 import Protected from './Protected';
 import CustomLogin from './CustomLogin';
@@ -40,7 +39,7 @@ const App: React.FC<{
         <Switch>
           <Route path='/login' component={CustomLogin}/>
           <Route path='/sessionToken-login' component={SessionTokenLogin}/>
-          <SecureRoute exact path='/protected' component={Protected} useOktaAuth={useOktaAuth} />
+          <SecureRoute exact path='/protected' component={Protected} />
           <Route path='/implicit/callback' component={LoginCallback} />
           <Route path='/pkce/callback' component={LoginCallback} />
           <Route path='/' component={Home}/>